+++ editors/vscode/src/syntax/illegal.ts
@@ -1,8 +1,0 @@
-export const illegal = {
-  captures: {
-    1: {
-      name: "invalid.illegal.toml",
-    },
-  },
-  match: "(.*)",
-};