+++ editors/vscode/src/syntax/index.ts
@@ -9,7 +9,7 @@ const syntax = {
   scopeName: "source.toml",
   uuid: "8b4e5008-c50d-11ea-a91b-54ee75aeeb97",
   information_for_contributors: [
-    "Originally was maintained by aster (galaster@foxmail.com). This notice is only kept here for the record, please don't send e-mails about problems and issues.",
+    "Originally was maintained by aster (galaster@foxmail.com). This notice is only kept here for the record, please don't send e-mails about bugs and other issues.",
   ],
   patterns: [
     {
@@ -34,7 +34,6 @@ const syntax = {
     table,
     entryBegin,
     value,
-    // illegal,
   },
 };